+++ sdk/aqueduct/artifact.py
@@ -3,6 +3,9 @@ import uuid
 from pydantic import BaseModel
 from typing import Any, Dict, Optional
 
+from aqueduct.enums import ArtifactType
+from aqueduct.error import AqueductError
+
 
 class ArtifactSpec(BaseModel):
     table: Optional[Dict[Any, Any]]
@@ -15,3 +18,15 @@ class Artifact(BaseModel):
     id: uuid.UUID
     name: str
     spec: ArtifactSpec
+
+def get_artifact_type(artifact: Artifact) -> ArtifactType:
+    if artifact.spec.table is not None:
+        return ArtifactType.TABLE
+    if artifact.spec.float is not None:
+        return ArtifactType.NUMBER
+    if artifact.spec.bool is not None:
+        return ArtifactType.BOOL
+    if artifact.spec.jsonable is not None:
+        return ArtifactType.PARAM
+    else:
+        raise AqueductError("Invalid operator type")
