+++ sdk/aqueduct/check_artifact.py
@@ -4,7 +4,7 @@ import json
 import uuid
 from typing import Optional, Dict, Any
 
-from aqueduct.utils import get_description_for_check
+from aqueduct.utils import get_description_for_check, format_header_for_print
 
 from aqueduct.api_client import APIClient
 from aqueduct.dag import DAG, apply_deltas_to_dag, SubgraphDAGDelta, UpdateParametersDelta
@@ -77,7 +77,6 @@ class CheckArtifact(Artifact):
 
     def describe(self) -> None:
         """Prints out a human-readable description of the check artifact."""
-        print("==================== CHECK ARTIFACT =============================")
         input_operator = self._dag.must_get_operator(with_output_artifact_id=self._artifact_id)
 
         general_dict = get_description_for_check(input_operator)
@@ -92,4 +91,5 @@ class CheckArtifact(Artifact):
             self._dag.must_get_artifact(artf).name for artf in input_operator.inputs
         ]
 
+        print(format_header_for_print(f"'{input_operator.name}' Check Artifact"))
         print(json.dumps(readable_dict, sort_keys=False, indent=4))
