+++ sdk/aqueduct/utils.py
@@ -32,6 +32,12 @@ API_KEY_HEADER = "api-key"
 CLIENT_VERSION = "sdk-client-version"
 
 
+# TODO: docstring
+def format_header_for_print(header: str) -> str:
+    prefix_len = 20
+    full_len = 80
+    return f"{'=' * prefix_len} {header} {'=' * max(0, full_len - prefix_len - len(header))}"
+
 def generate_auth_headers(api_key: str) -> Dict[str, str]:
     return {API_KEY_HEADER: api_key, CLIENT_VERSION: str(__version__)}
